+++ premier-design/__mocks__/styled-system.js
@@ -1,0 +1,3 @@
+export const css = () => '';
+export const cva = () => () => '';
+export const cx = (...args) => args.filter(Boolean).join(' ');